server/requirements.txt
# Smart Campus Assistant — Python dependencies
# Pin minor versions for reproducible builds; patch bumps are safe.

fastapi==0.111.*
uvicorn[standard]==0.29.*
pydantic==2.*

# Database (Task 3.2)
sqlalchemy==2.0.*
# AI client (Task 3.4)
openai==1.48.*
python-dotenv==1.0.*

# Testing
pytest==7.4.*
httpx==0.27.*
